Move cli-related scripts to cli subfolder. Update requirements.txt
