+++ credentialdigger/__init__.py
@@ -1,6 +1,3 @@
 from .client import Client
 from .client_postgres import PgClient
 from .client_sqlite import SqliteClient
-from .add_rules import add_rules
-from .download import download
-from .scan import scan
