+++ credentialdigger/cli/add_rules.py
