+++ credentialdigger/cli/scan.py
